etl/requirements.txt
pandas==2.2.2
python-dotenv==1.0.1
psycopg[binary]==3.2.1
SQLAlchemy==2.0.35
